+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Implemented `Borrow` and `BorrowMut` for `DivBuf` and `DivBufMut`
 
 ### Changed
+- Futures should do nothing until polled.
+  ([#2](https://github.com/asomers/futures-locks/issues/2))
 - Fixed potential deadlocks when dropping Futures without first polling them.
   ([#1](https://github.com/asomers/futures-locks/issues/1))